Move logging related classes into own directory.
