+++ lib/vedeu/all.rb
@@ -2,8 +2,9 @@ require 'vedeu/version'
 require 'vedeu/runtime/launcher'
 require 'vedeu/runtime/bootstrap'
 require 'vedeu/exceptions'
-require 'vedeu/log'
-require 'vedeu/debug'
+require 'vedeu/log/log'
+require 'vedeu/log/debug'
+require 'vedeu/log/timer'
 require 'vedeu/traps'
 require 'vedeu/common'
 require 'vedeu/terminal_mode'
@@ -14,7 +15,6 @@ require 'vedeu/configuration/api'
 require 'vedeu/configuration/configuration'
 
 require 'vedeu/terminal'
-require 'vedeu/timer'
 require 'vedeu/runtime/main_loop'
 require 'vedeu/runtime/application'
