+++ lib/vedeu/log/log.rb
