+++ lib/vedeu/log/timer.rb
