+++ test/lib/vedeu/internal_api_test.rb
@@ -8,6 +8,7 @@ module Vedeu
     it { Vedeu.must_respond_to(:borders) }
     it { Vedeu.must_respond_to(:buffers) }
     it { Vedeu.must_respond_to(:cursors) }
+    it { Vedeu.must_respond_to(:debug) }
     it { Vedeu.must_respond_to(:events) }
     it { Vedeu.must_respond_to(:foreground_colours) }
     it { Vedeu.must_respond_to(:geometries) }
