+++ test/lib/vedeu/log/debug_test.rb
@@ -1,0 +1,16 @@
+require 'test_helper'
+
+module Vedeu
+
+  describe Debug do
+
+    let(:described) { Vedeu::Debug }
+
+    describe '.debug' do
+      # @todo
+      # it { skip }
+    end
+
+  end # Debug
+
+end # Vedeu
