+++ test/lib/vedeu/log/log_test.rb
