+++ test/lib/vedeu/log/timer_test.rb
